requirements.txt
Django==4.2.7 
djangorestframework==3.14.0 
django-cors-headers==4.3.0 
psycopg2-binary==2.9.9 
python-decouple==3.8 
Pillow==10.1.0 
pytest==7.4.3 
pytest-django==4.7.0 
requests==2.31.0 
groq==1.0.0 
celery==5.6.2 
redis==7.0.1
django-celery-beat==2.9.0
djangorestframework-simplejwt==5.5.1